import: Move 'build_avatar' to import_util.
